audioEffectManager, playbackManager, systemSoundsManager: silence the stragglers
Drop the unused returnValue in _getAudioEffectsStatus and the unused
LSSubscriptionNext() capture in notifyGetPlayabackStatus (the call itself
must stay - it advances the iterator that LSSubscriptionRemove acts on).
In _playFeedback, compare the snprintf length against the size_t buffer
size through an explicit cast (-Wsign-compare); the negative-result case
is already handled, or falls out as a huge unsigned value that fails the
bound check just as before.

+++ src/modules/systemSoundsManager/systemSoundsManager.cpp
@@ -96,7 +96,7 @@ bool SystemSoundsManager::_playFeedback(LSHandle *lshandle, LSMessage *message, 
         PM_LOG_ERROR(MSGID_SYSTEMSOUND_MANAGER, INIT_KVCOUNT,"snprintf failed to generate filename. Returning from here\n");
         goto error;
     }
-    else if (snprintf(filename, size, SYSTEMSOUNDS_PATH "%s-ondemand.pcm", name.c_str()) >= size)
+    else if ((size_t)snprintf(filename, size, SYSTEMSOUNDS_PATH "%s-ondemand.pcm", name.c_str()) >= size)
     {
         PM_LOG_ERROR(MSGID_SYSTEMSOUND_MANAGER, INIT_KVCOUNT,"snprintf buffer overflow. Filename might be truncated\n");
     }
@@ -109,7 +109,7 @@ bool SystemSoundsManager::_playFeedback(LSHandle *lshandle, LSMessage *message, 
          * no -ondemand suffix, including the AdjustVolume one the volume keys use,
          * and rejecting those here is what made the keys silent.
          */
-        if (snprintf(filename, size, SYSTEMSOUNDS_PATH "%s.pcm", name.c_str()) < size)
+        if ((size_t)snprintf(filename, size, SYSTEMSOUNDS_PATH "%s.pcm", name.c_str()) < size)
             fp = fopen(filename, "r");
     }
     free(filename);
